big_brain/models/vision/ml-fastvlm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0 